code cleanup, follow one class per file rule and avoid too many callbacks
