+++ src/BannerNameCommand.cs
@@ -4,15 +4,8 @@ using ChatCommands;
 
 namespace ScarabolMods
 {
-  [ModLoader.ModManager]
   public class BannerNameChatCommand : IChatCommand
   {
-    [ModLoader.ModCallback (ModLoader.EModCallbackType.AfterItemTypesDefined, "scarabol.commands.bannername.registercommand")]
-    public static void AfterItemTypesDefined ()
-    {
-      CommandManager.RegisterCommand (new BannerNameChatCommand ());
-    }
-
     public bool IsCommand (string chat)
     {
       return chat.Equals ("/bannername");
