+++ src/GiveAllCommand.cs
@@ -6,14 +6,8 @@ using Permissions;
 
 namespace ScarabolMods
 {
-  [ModLoader.ModManager]
   public class GiveAllChatCommand : IChatCommand
   {
-    [ModLoader.ModCallback (ModLoader.EModCallbackType.AfterItemTypesDefined, "scarabol.commands..registercommand")]
-    public static void AfterItemTypesDefined ()
-    {
-      CommandManager.RegisterCommand (new GiveAllChatCommand ());
-    }
 
     public bool IsCommand (string chat)
     {
@@ -22,7 +16,7 @@ namespace ScarabolMods
 
     public bool TryDoCommand (Players.Player causedBy, string chattext)
     {
-      if (!PermissionsManager.CheckAndWarnPermission (causedBy, CommandsModEntries.MOD_PREFIX + "giveall")) {
+      if (!PermissionsManager.CheckAndWarnPermission (causedBy, AntiGrief.MOD_PREFIX + "giveall")) {
         return true;
       }
       var m = Regex.Match (chattext, @"/giveall (?<material>.+) (?<amount>\d+)");
