+++ src/KillNPCsCommand.cs
@@ -7,14 +7,9 @@ using NPC;
 
 namespace ScarabolMods
 {
-  [ModLoader.ModManager]
+
   public class KillNPCsChatCommand : IChatCommand
   {
-    [ModLoader.ModCallback (ModLoader.EModCallbackType.AfterItemTypesDefined, "scarabol.commands.killnpcs.registercommand")]
-    public static void AfterItemTypesDefined ()
-    {
-      CommandManager.RegisterCommand (new KillNPCsChatCommand ());
-    }
 
     public bool IsCommand (string chat)
     {
@@ -23,7 +18,7 @@ namespace ScarabolMods
 
     public bool TryDoCommand (Players.Player causedBy, string chattext)
     {
-      if (!PermissionsManager.CheckAndWarnPermission (causedBy, CommandsModEntries.MOD_PREFIX + "killnpcs")) {
+      if (!PermissionsManager.CheckAndWarnPermission (causedBy, AntiGrief.MOD_PREFIX + "killnpcs")) {
         return true;
       }
       var m = Regex.Match (chattext, @"/killnpcs (?<targetplayername>['].+?[']|[^ ]+)");
