+++ src/ServerPopCommand.cs
@@ -3,14 +3,9 @@ using ChatCommands;
 
 namespace ScarabolMods
 {
-  [ModLoader.ModManager]
+
   public class ServerPopCommand : IChatCommand
   {
-    [ModLoader.ModCallback (ModLoader.EModCallbackType.AfterItemTypesDefined, "scarabol.commands.serverpopcommand.registercommand")]
-    public static void AfterItemTypesDefined ()
-    {
-      CommandManager.RegisterCommand (new ServerPopCommand ());
-    }
 
     public bool IsCommand (string chat)
     {
