+++ src/WarpPlaceCommand.cs
@@ -7,14 +7,9 @@ using ChatCommands.Implementations;
 
 namespace ScarabolMods
 {
-  [ModLoader.ModManager]
+
   public class WarpPlaceChatCommand : IChatCommand
   {
-    [ModLoader.ModCallback (ModLoader.EModCallbackType.AfterItemTypesDefined, "scarabol.commands.warpplace.registercommand")]
-    public static void AfterItemTypesDefined ()
-    {
-      CommandManager.RegisterCommand (new WarpPlaceChatCommand ());
-    }
 
     public bool IsCommand (string chat)
     {
@@ -23,7 +18,7 @@ namespace ScarabolMods
 
     public bool TryDoCommand (Players.Player causedBy, string chattext)
     {
-      if (!PermissionsManager.CheckAndWarnPermission (causedBy, CommandsModEntries.MOD_PREFIX + "warp.place")) {
+      if (!PermissionsManager.CheckAndWarnPermission (causedBy, AntiGrief.MOD_PREFIX + "warp.place")) {
         return true;
       }
       var m = Regex.Match (chattext, @"/warpplace (?<px>-?\d+) (?<py>-?\d+)( (?<pz>-?\d+))?");
